+++ src/dbally/audit/__init__.py
@@ -7,9 +7,9 @@ try:
 except ImportError:
     pass
 
-from .event_span import EventSpan
 from .event_tracker import EventTracker
 from .events import LLMEvent, RequestEnd, RequestStart, SimilarityEvent
+from .spans import EventSpan
 
 __all__ = [
     "CLIEventHandler",
